refactor: random fixes, import resolution
This commit addresses a few things.
It tries and addresses imports that might not be resolved
completely such as cv2. A workaround in future might be
creating a requirements.txt file and asking users to install that
instead. The 'hack' that I've implemented installs the requirements using subprocess,
however it's considered a hack because it assumes that python is a path-wide
environment variable that the user has set.

Author: Cyré